更改百度地图数据显隐方式 然而还是很慢 高德地图数据传输添加

+++ app/src/main/java/com/xunce/gsmr/util/gps/MapHelper.java
@@ -34,7 +34,7 @@ public class MapHelper {
      * @param baiduMap
      * @param zoomLevel
      */
-    public static void animateZoom(BaiduMap baiduMap, int zoomLevel) {
+    public static void animateZoom(BaiduMap baiduMap, float zoomLevel) {
         MapStatusUpdate u = MapStatusUpdateFactory.zoomTo(zoomLevel);
         baiduMap.animateMapStatus(u);
     }
